Do refactor in NetworkManager and split  into LZ4Utils and PlayerHandle
